requirements.txt
asgiref==3.5.0
Django==4.0.4
sqlparse==0.4.2
psycopg2-binary==2.9.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
python-dotenv==0.20.0
gunicorn==20.1.0
